refactor(app): extract session active message controller
